Follow the 'connections' uri when present.
* Full access to the API requires both user consent and designating
which organizations the user wants the application to be able to
interact with. Follow the 'connections' uri if present on any org to
allow the user finish the setup process.
* Re-worked the `/` controller slightly. Previously it was always
updating the session `settings` with a new set of values from the oauth
configuration. This caused the access token in the session to be lost
when coming back to the application after the connections org-select
redirect. Now it only initializes session `settings` if it is not set.